refactor(#2070): consolidate close-at-or-before reads into one Decimal-contract helper
close_row_at_or_before moves to thesis_dq_audit (import direction:
thesis_outcomes already depends on dq_audit). dq_audit delegates and
converts via _to_float at the call site; find_stale's mint-close LATERAL
cites the helper as the SQL-side mirror. One Python implementation
remains; Decimal end-to-end contract (PR #2061 review) now covers both
readers.

Closes #2070
