+++ app/services/thesis_outcomes.py
@@ -29,13 +29,12 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 from datetime import date, timedelta
-from decimal import Decimal
 from typing import Any
 
 import psycopg
 import psycopg.rows
 
-from app.services.thesis_dq_audit import _anchor_block, _parse_iso_date
+from app.services.thesis_dq_audit import _anchor_block, _parse_iso_date, close_row_at_or_before
 
 HORIZONS: tuple[int, ...] = (30, 90, 365)
 
@@ -117,36 +116,6 @@ def classify_immature(
     if (due_date - max_price_date).days > _SERIES_DEAD_GRACE_DAYS:
         return "series_dead"
     return "immature_series_stalled"
-
-
-def close_row_at_or_before(
-    conn: psycopg.Connection[Any], instrument_id: int, as_of: date
-) -> tuple[date, Decimal] | None:
-    """(price_date, close) of the last print at or before ``as_of``.
-
-    Same read as ``thesis_dq_audit._close_at_or_before`` but the ledger
-    also needs the trading day actually used (``realized_date``), hence
-    the widened return shape rather than a shared helper. The close stays
-    ``Decimal`` end-to-end (NUMERIC in, NUMERIC out — no binary-float
-    round-trip; PR #2061 review); a non-finite NUMERIC (NaN) maps to None
-    the same way the ``_to_float`` chokepoint would.
-    """
-    with conn.cursor() as cur:
-        cur.execute(
-            """
-            SELECT price_date, close FROM price_daily
-            WHERE instrument_id = %(iid)s AND price_date <= %(as_of)s
-            ORDER BY price_date DESC LIMIT 1
-            """,
-            {"iid": instrument_id, "as_of": as_of},
-        )
-        row = cur.fetchone()
-    if row is None:
-        return None
-    close = row[1]
-    if not isinstance(close, Decimal) or not close.is_finite():
-        return None
-    return (row[0], close)
 
 
 # All thesis versions with at least one horizon uncaptured. The run
